Add value field to Button element
When building an interactive slack bot it can be useful to have a button
trigger an action on the backend but pass along some special ID that the
backend can then use to lookup something useful.

E.g. Having a feedback button that you want to pass along a parameter to
identify what you're giving feedback on.

https://api.slack.com/reference/block-kit/block-elements#button__fields